Let's use primitive types that will not force us in having hard checks <= 0, since the compiler will enforce them automatically.

+++ src/SoundFingerprinting.Tests/Integration/SubFingerprintDaoTest.cs
@@ -12,7 +12,6 @@
     using SoundFingerprinting.DAO.Data;
     using SoundFingerprinting.Data;
     using SoundFingerprinting.InMemory;
-    using SoundFingerprinting.Strides;
 
     [TestFixture]
     [Category("RequiresWindowsDLL")]
@@ -45,8 +44,8 @@
                             new HashedFingerprint(
                                 GenericSignature(),
                                 genericHashBuckets,
-                                sequenceNumber,
-                                sequenceNumber * 0.928,
+                                (uint)sequenceNumber,
+                                sequenceNumber * 0.928f,
                                 Enumerable.Empty<string>()));
 
             InsertHashedFingerprintsForTrack(hashedFingerprints, trackReference);
